Update so that when new game clicked takes player to board
