+++ vue-frontend/src/pages/NewGame.vue
@@ -10,21 +10,6 @@
           <b-button size="lg" @click="handleClick()">Start New Game</b-button>
         </b-col>
       </b-row>
-      <b-row 
-        v-if="shareableUrl">
-        <b-col cols="6" offset="3">
-          <b-input-group prepend="Share with your opponent" class="mt-3">
-          <b-form-input id="nextPlayerUrl" v-model="shareableUrl"></b-form-input>
-          <b-input-group-append>
-            <button type="button" @click="handleCopy">
-              <span class="far fa-copy" />
-            </button>
-          </b-input-group-append>
-        </b-input-group>
-          <p v-if="copySucceeded">Copied!</p>
-          <p v-else>Press CTRL+C to copy.</p>
-        </b-col>
-      </b-row>
   </Section>
 </template>
 
@@ -35,8 +20,6 @@ export default {
   name: 'NewGamePage',
   data: function() {
     return {
-      shareableUrl: null,
-      copySucceeded: null
     }
   },
   methods: {
@@ -46,24 +29,11 @@ export default {
             this.copySucceeded = false
             this.gameId = response.data.gameId;
             this.playerId = response.data.currentPlayer;
-            this.nextPlayerId = this.generateNextPlayer(this.playerId);
-            this.shareableUrl = `${window.location.origin}/${this.gameId}/${this.nextPlayerId}`;
             console.log(response);
+            this.$router.push(`${this.gameId}/${this.playerId}`)
           } catch (error) {
             console.error(error);
           }
-    },
-    handleCopy: function () {
-        this.$copyText(this.shareableUrl)
-        this.copySucceeded = true
-      },
-    generateNextPlayer(playerId){
-      if (playerId == 'r') {
-        return 'y'
-      }
-      else if (playerId === 'y') {
-        return 'r'
-      }
     }
   },
